app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib
bcrypt
python-jose[cryptography]
httpx
gunicorn
email-validator
python-multipart